refactor(agent): split agent.go into agent, loop, history
